requirements.txt
streamlit>=1.21.0
numpy>=1.20.0
matplotlib>=3.5.0
opencv-python>=4.5.0
scipy>=1.7.0
pygam>=0.8.0
pillow>=8.0.0
bresenham>=0.2.0
scikit-image>=0.19.0
seaborn>=0.11.0
pandas>=1.3.0
